Fixed TBA verification handling
The reason it didn't work before was because TBA uses python's json.dumps() to generate the HMAC hash, but on JS we have to use JSON.stringify(), which removes spaces from the final string, something that dumps() doesn't do. Found a StackOverflow answer to replicate dumps' behavior and cited it.